requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aioresponses==0.7.4
aleph_alpha_client==3.2.0
apify_client==1.3.1
asyncer==0.0.2
atlassian==0.0.0
auth0_python==4.4.0
azure_storage==0.37.0
beautifulsoup4==4.12.2
bibtexparser==1.4.0
bilibili_api==9.1.0
boto3==1.26.159
chainlit==0.6.1
chromadb==0.3.26
click==8.1.3
couchdb3==1.2.0
databricks==0.2
dataclasses_json==0.5.8
deepspeed==0.10.0
diskcache==5.6.1
docarray==0.37.0
docx2txt==0.8
duckduckgo_search==3.8.4
ebooklib==0.18
fastapi==0.101.0
fastapi_socketio==0.0.10
feedly==0.11.3
feedparser==6.0.10
filetype==1.2.0
firebase_admin==6.2.0
fitz==0.0.1.dev2
geopy==2.3.0
gkeepapi==0.14.2
google_api_python_client==2.95.0
google_auth_oauthlib==1.0.0
googlemaps==4.10.0
gptcache==0.1.36
gql==3.4.1
html2text==2020.1.16
httpx==0.24.1
huggingface_hub==0.16.4
InstructorEmbedding==1.0.1
ipykernel==6.25.0
ipython==8.14.0
jsonschema==4.18.6
langchain==0.0.253
llama_hub==0.0.18
lxml==4.9.3
matplotlib==3.7.2
metal_sdk==2.0.1
modelscope==1.8.1
msal==1.22.0
nbconvert==7.7.3
neo4j==5.11.0
nest_asyncio==1.5.6
nltk==3.8.1
numpy==1.25.1
odps==3.5.1
openapi_schema_pydantic==1.2.4
opendal==0.39.0
osmxtract==0.0.1
pandas==2.0.2
pandasai==0.8.2
pdf2image==1.16.3
pdfminer==20191125
pexpect==4.8.0
Pillow==9.5.0
Pillow==10.0.0
pinecone_text==0.5.2
platformdirs==3.10.0
playwright==1.36.0
praw==7.7.1
protobuf==4.23.4
pyairtable==2.0.0
pydantic==1.10.11
pydrive==1.3.1
pydub==0.25.1
pygtrie==2.5.0
pymilvus==2.2.14
pymongo==4.4.1
pymysql==1.1.0
pyowm==3.3.0
pyparsing==3.1.1
pypdf==3.15.0
pyspark==3.4.1
pytesseract==0.3.10
pytest==7.4.0
python-dotenv==1.0.0
python-socketio==5.8.0
python_graphql_client==0.4.3
PyYAML==6.0
PyYAML==6.0.1
qdrant_client==1.4.0
rdflib==7.0.0
regex==2023.6.3
reportlab==4.0.4
responses==0.23.3
retrying==1.3.4
scikit_learn==1.3.0
sentence_transformers==2.2.2
sentencepiece==0.1.99
slack_sdk==3.21.3
snscrape==0.7.0.20230622
spacy==3.6.0
spotipy==2.23.0
SQLAlchemy==2.0.19
starlette==0.31.0
svglib==1.5.1
syncer==2.0.3
tenacity==8.2.2
tensorflow_hub==0.14.0
tensorflow_text==2.13.0
tiktoken==0.4.0
tomli==2.0.1
torch==2.0.1
tqdm==4.65.0
trafilatura==1.6.1
tweepy==4.14.0
typing_extensions==4.7.1
uptrace==1.19.0
uvicorn==0.23.2
watchfiles==0.19.0
whisper==1.1.10
xlrd==2.0.1
yelpapi==2.5.0
youtube_search==2.1.2
youtube_transcript_api==0.6.1
zep_python==0.34
